requirements.txt
pytesseract~=0.3.3
pymystem3~=0.2.0
Pillow~=7.1.1
pandas~=1.0.3
sklearn~=0.0
scikit-learn~=0.22.2.post1
Flask-RESTful==0.3.8
numpy~=1.18.2
seaborn~=0.10.0
matplotlib~=3.2.1
scipy~=1.4.1
Flask~=1.1.2
gunicorn